improved stability in nl-vl

+++ src/post/mod_post_chordwise.f90
@@ -152,7 +152,7 @@ subroutine post_chordwise(sbprms, basename, data_basename, an_name, &
   real(wp), allocatable                                   :: span_station(:)
   real(wp), allocatable                                   :: y_cen_tras(:) 
   real(wp), allocatable                                   :: chord(:) 
-  real(wp)                                                :: chord_start, chord_end
+  integer                                                 :: chord_start, chord_end
   integer,  allocatable                                   :: id_minus(:), id_plus(:) 
   real(wp)                                                :: u_inf(3), P_inf, rho 
   !> field to interpolate  
@@ -310,7 +310,7 @@ subroutine post_chordwise(sbprms, basename, data_basename, an_name, &
               abs ( comps(id_comp)%loc_points(ax_coor,comps(id_comp)%el(ie)%i_ver(1) )&
               - comps(id_comp)%loc_points(ax_coor,comps(id_comp)%el(ie)%i_ver(2) ) )
       !> local chord as projection of the profile on x-z plane
-      chord(is) = sqrt((abs(minval(comps(id_comp)%loc_points(ax_coor - 1,chord_start:chord_end))) + & 
+      chord(is) = sqrt((abs(minval(comps(id_comp)%loc_points(ax_coor - 1 ,chord_start:chord_end))) + & 
               abs(maxval(comps(id_comp)%loc_points(ax_coor - 1,chord_start:chord_end))))**2 + &
               (abs(minval(comps(id_comp)%loc_points(ax_coor + 1,chord_start:chord_end))) + & 
               abs(maxval(comps(id_comp)%loc_points(ax_coor + 1,chord_start:chord_end))))**2)
